requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0

SQLAlchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.10
alembic==1.13.3

python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.2
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

redis==5.0.4

httpx==0.27.2
pytest==8.3.3
